# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a REST API for the RyftOne Server and Cluster. It provides methods to list files and folders stored on server or cluster and execute search and count requests. Both RAW_TEXT and RECORD searches supported. Please explore API below for details and examples.",
    "title": "RyftOne REST API",
    "version": "0.12.0"
  },
  "paths": {
...
    "/files": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "List a directory content or download a file",
        "description": "Endpoint to query node/cluster file system structure or download a file",
        "operationId": "getFiles",
        "parameters": [
          {
            "name": "dir",
            "type": "string",
            "description": "The folder path to return child items of. Default '/'",
            "in": "query"
          },
          {
            "name": "hidden",
            "type": "boolean",
            "description": "Report hidden files",
            "in": "query"
          },
          {
            "name": "local",
            "type": "boolean",
            "description": "Return file structure from local node only or whole cluster",
            "in": "query"
          },
          {
            "name": "catalog",
            "type": "string",
            "description": "The path to catalog to download file of",
            "in": "query"
          },
          {
            "name": "file",
            "type": "string",
            "description": "The filepath to download",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of files and directories for node/cluster",
            "schema": {
              "$ref": "#/definitions/GetFilesResponse"
            }
          },
          "default": {
            "description": "Error responses",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload files",
        "description": "Endpoint to upload files",
        "operationId": "postFiles",
        "consumes": [
          "application/octet-stream",
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "content",
            "in": "formData",
            "type": "file",
            "description": "File content to upload",
            "required": true
          },
          {
            "name": "file",
            "type": "string",
            "description": "The uploading filename. Use {{random}} for random names.",
            "required": true,
            "in": "query"
          },
          {
            "name": "catalog",
            "type": "string",
            "description": "The catalog to append file to.",
            "in": "query"
          },
          {
            "in": "query",
            "type": "string",
            "name": "delimiter",
            "description": "The delimiter is used between file parts in catalog."
          },
          {
            "name": "offset",
            "type": "integer",
            "format": "int64",
            "description": "The optional file part offset.",
            "in": "query"
          },
          {
            "name": "length",
            "type": "integer",
            "format": "int64",
            "description": "The optional file part length.",
            "in": "query"
          },
          {
            "in": "query",
            "type": "string",
            "name": "lifetime",
            "description": "The file or catalog lifetime, for example \"2h\"."
          },
          {
            "in": "query",
            "type": "string",
            "name": "share-mode",
            "x-go-name": "ShareMode",
            "description": "Sharing mode used to access data files. Can be \"wait-10s\" or \"ignore\". By default \"wait-0s\" is used."
          },
          {
            "in": "query",
            "type": "boolean",
            "name": "local",
            "description": "Perform action only on this cluster member. Default 'false'."
          }
        ],
        "responses": {
          "200": {
            "description": "Map of files and directories uploaded",
            "schema": {
              "$ref": "#/definitions/PostFilesResponse"
            }
          },
          "default": {
            "description": "Error responses",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "files"
        ],
        "summary": "Delete directories and files",
        "description": "Endpoint to delete node's directories and files",
        "operationId": "deleteFiles",
        "parameters": [
          {
            "name": "dir",
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "The directory path to delete. Wildcards are supported.",
            "in": "query"
          },
          {
            "name": "file",
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "The file path to delete. Wildcards are supported.",
            "in": "query"
          },
          {
            "name": "catalog",
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi",
            "description": "The catalog path to delete. Wildcards are supported.",
            "in": "query"
          },
          {
            "in": "query",
            "type": "boolean",
            "name": "local",
            "x-go-name": "Local",
            "description": "Perform action only on this cluster member. Default 'false'."
          }
        ],
        "responses": {
          "200": {
            "description": "Map of files and directories deleted",
            "schema": {
              "$ref": "#/definitions/DeleteFilesResponse"
            }
          },
          "default": {
            "description": "Error responses",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "additionalProperties": {
        "type": "object"
      },
      "properties": {
        "message": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      }
    },
...
    "PostFilesResponse": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "properties": {
              "length": {
                "type": "integer"
              },
              "offset": {
                "type": "integer"
              },
              "path": {
                "type": "string"
              },
              "error": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "DeleteFilesResponse": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "properties": {
              "path": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "GetFilesResponse": {
      "type":"string"
    }
  },
  "securityDefinitions": {
    "jwt": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "basic": [],
      "jwt": []
    }
  ]
}